requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11